Add files via upload
This Matlab script is written in order to produce the time dependent values of average power during and at the end of an ON period given the values of t are less than the values of tau_ON. When the script is ran, its important to look at the values being produced and how they correspond to one another and the length of time of the ON period. As well as how they compare a non time varying situation. One thing to consider are the values of omega_d and omega_dprime; The value of Omega_d can be found using the equation SQRT((1/2*tau)^2 - (omega_o)^2). Since the value of 1/2*tau is so small when the equation is carried out, I was curious as if it was negligible in this situation? Same thing for omega_dprime except with tau_prime instead.